refactor: shared unique name validator
