+++ backend/apps/post/models.py
@@ -5,9 +5,9 @@ from dynamic_filenames import FilePattern
 
 from apps.community.models import Community
 from apps.user.models import Profile
-from mixins.models.created_at import CreatedAtMixin
-from mixins.models.shortuuid import ShortUUIDMixin
-from mixins.models.type import TypeMixin
+from shared.mixins.models.created_at import CreatedAtMixin
+from shared.mixins.models.shortuuid import ShortUUIDMixin
+from shared.mixins.models.type import TypeMixin
 
 from .managers import PostManager
