+++ backend/shared/mixins/models/avatar.py
